refactor: IssueDetail 폴더 구조 정리
- ghojeong/issue-tracker/#183
