+++ frontend/issue-cracker/src/components/layout/IssueDetail/IssueDetailHeader/IssueDetailHeaderInfoSection/IssueDetailBadge.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { TYPE as T } from '../../../utils/const';
-import LabelLargeGroup from '../../common/group/LabelLargeGroup';
+import { TYPE as T } from '../../../../../utils/const';
+import LabelLargeGroup from '../../../../common/group/LabelLargeGroup';
 
 const IssueDetailBadge = ({ state }: { state: boolean }): JSX.Element => {
   return (
